[evals] Keep trace chat format with trace processor
